Andy

+++ src/services/commerce/product.ts
@@ -4,9 +4,10 @@ import { SUCCESS, CREATED, SERVER_ERROR, BAD_REQUEST } from '../../utils/statusC
 import { CreateProductIProps } from '../../types/product';
 
 export const createProduct = async (req: Request, res: Response) => {
-    const { name, category, price, description, address, city, state, contact, image, userId, storeId } = req.body;
+    const { name, category, price, description, address, city, state, contact, image, storeId, draft, inStock } = req.body;
+    const userId = req.user.userId
     const newProduct: CreateProductIProps = {
-        storeId, name, category, price, description, address, city, state, contact, image, userId
+        storeId, name, category, price, description, address, city, state, contact, image, userId, draft, inStock
     }
     try {
         const result = await Product.create(newProduct)
@@ -19,14 +20,14 @@ export const createProduct = async (req: Request, res: Response) => {
 }
 
 export const updateProduct = (req: Request, res: Response) => {
-    const { productId } = req.params
-    const { name, category, price, description, address, city, state, contact, image, userId, storeId } = req.body;
+    const { name, category, price, description, address, city, state, contact, image, draft, inStock, productId } = req.body;
+    const userId = req.user.userId
     const editedProduct: CreateProductIProps = {
-        storeId, name, category, price, description, address, city, state, contact, image, userId
+        name, category, price, description, address, city, state, contact, image, userId, draft, inStock
     }
-    Product.findOneAndUpdate({ _id: productId }, editedProduct, async (err: unknown, data: unknown) => {
+    Product.findOneAndUpdate({ _id: productId, userId: userId }, editedProduct, async (err: unknown, data: unknown) => {
         if (data === null || !data) {
-            return res.status(BAD_REQUEST).json({ status: BAD_REQUEST, message: `product id ${productId} not found`, success: false });
+            return res.status(BAD_REQUEST).json({ status: BAD_REQUEST, message: `product id ${productId} not found or access denied`, success: false });
         } else if (err) {
             return res.status(SERVER_ERROR).json({ status: SERVER_ERROR, message: err, success: false });
         } else {
@@ -36,36 +37,13 @@ export const updateProduct = (req: Request, res: Response) => {
     })
 }
 
-export const getAllProducts = async (req: Request, res: Response) => {
-    const { storeId } = req.params
-    try {
-        const result = await Product.find({ storeId })
-        if (result) {
-            return res.status(SUCCESS).json({ status: SUCCESS, message: 'ok', success: true, payload: result });
-        }
-    } catch (error) {
-        return res.status(SERVER_ERROR).json({ status: SERVER_ERROR, message: error, success: false });
-    }
-}
-
-export const getSingleProductById = async (req: Request, res: Response) => {
-    const { productId } = req.params
-    Product.findById({ _id: productId }, async (err: unknown, data: unknown) => {
-        if (data === null || !data) {
-            return res.status(BAD_REQUEST).json({ status: BAD_REQUEST, message: `product id ${productId} not found`, success: false });
-        } else if (err) {
-            return res.status(SERVER_ERROR).json({ status: SERVER_ERROR, message: err, success: false });
-        } else {
-            return res.status(SUCCESS).json({ status: SUCCESS, message: 'Product updated', success: true, payload: data});
-        }
-    })
-}
 
 export const deleteProduct = (req: Request, res: Response) => {
     const { productId } = req.params
-    Product.findOneAndDelete({ _id: productId }, async (err: unknown, data: unknown) => {
+    const userId = req.user.userId
+    Product.findOneAndDelete({ _id: productId, userId: userId }, async (err: unknown, data: unknown) => {
         if (data === null || !data) {
-            return res.status(BAD_REQUEST).json({ status: BAD_REQUEST, message: `product id ${productId} not found`, success: false });
+            return res.status(BAD_REQUEST).json({ status: BAD_REQUEST, message: `product id ${productId} not found or access denied`, success: false });
         } else if (err) {
             return res.status(SERVER_ERROR).json({ status: SERVER_ERROR, message: err, success: false });
         } else {
@@ -74,6 +52,96 @@ export const deleteProduct = (req: Request, res: Response) => {
         }
     })
 }
+
+
+export const setProductStockStatus = async (req: Request, res: Response): Promise<Response> => {
+    const { productId, inStock } = req.body
+    const userId = req.user.userId
+    try {
+        const product = await Product.findOne({ _id: productId, userId: userId })
+        if (!product) {
+            return res.status(BAD_REQUEST).json({ status: BAD_REQUEST, message: `No product found for ${productId} or access denied`, success: false });
+        }
+        const updatedProduct = await Product.findByIdAndUpdate({ _id: productId }, { inStock: inStock })
+        if (updatedProduct) {
+            return res.status(CREATED).json({ status: CREATED, message: 'Product in-stock updated', success: true });
+        }
+    } catch (error) {
+        return res.status(SERVER_ERROR).json({ status: SERVER_ERROR, message: error, success: false });
+    }
+}
+
+
+export const setProductDraft = async (req: Request, res: Response): Promise<Response> => {
+    const { productId, draft } = req.body
+    const userId = req.user.userId
+    try {
+        const product = await Product.findOne({ _id: productId, userId: userId })
+        if (!product) {
+            return res.status(BAD_REQUEST).json({ status: BAD_REQUEST, message: `No product found for ${productId} or access denied`, success: false });
+        }
+        const updatedProduct = await Product.findByIdAndUpdate({ _id: productId }, { draft: draft })
+        if (updatedProduct) {
+            return res.status(CREATED).json({ status: CREATED, message: 'Product draft updated', success: true });
+        }
+    } catch (error) {
+        return res.status(SERVER_ERROR).json({ status: SERVER_ERROR, message: error, success: false });
+    }
+}
+
+
+export const getAllProductsVendor = async (req: Request, res: Response) => {
+    const { storeId } = req.params
+    const userId = req.user.userId
+    Product.find({ storeId: storeId, userId: userId }, async (err: unknown, data: unknown) => {
+        if (data === null || !data) {
+            return res.status(BAD_REQUEST).json({ status: BAD_REQUEST, message: `store id ${storeId} not found or access denied`, success: false });
+        } else if (err) {
+            return res.status(SERVER_ERROR).json({ status: SERVER_ERROR, message: err, success: false });
+        } else {
+            return res.status(SUCCESS).json({ status: SUCCESS, message: 'ok', success: true, payload: data });
+        }
+    })
+}
+
+
+export const getAllProductUser = async (req: Request, res: Response): Promise<Response> => {
+    const { storeId } = req.params
+    try {
+        const result = await Product.find({ storeId: storeId, draft: false })
+        return res.status(SUCCESS).json({ status: SUCCESS, message: 'ok', success: true, payload: result });
+    } catch (error) {
+        return res.status(SERVER_ERROR).json({ status: SERVER_ERROR, message: error, success: false });
+    }
+}
+
+export const getSingleProductByIdVendor = async (req: Request, res: Response) => {
+    const { productId } = req.params
+    const userId = req.user.userId
+    Product.findOne({ _id: productId, userId: userId }, async (err: unknown, data: unknown) => {
+        if (data === null || !data) {
+            return res.status(BAD_REQUEST).json({ status: BAD_REQUEST, message: `product id ${productId} not found`, success: false });
+        } else if (err) {
+            return res.status(SERVER_ERROR).json({ status: SERVER_ERROR, message: err, success: false });
+        } else {
+            return res.status(SUCCESS).json({ status: SUCCESS, message: 'Product updated', success: true, payload: data });
+        }
+    })
+}
+
+export const getSingleProductByIdUser = async (req: Request, res: Response) => {
+    const { productId } = req.params
+    Product.findOne({ _id: productId, draft: false }, async (err: unknown, data: unknown) => {
+        if (data === null || !data) {
+            return res.status(BAD_REQUEST).json({ status: BAD_REQUEST, message: `product id ${productId} not found or product not available but in draft`, success: false });
+        } else if (err) {
+            return res.status(SERVER_ERROR).json({ status: SERVER_ERROR, message: err, success: false });
+        } else {
+            return res.status(SUCCESS).json({ status: SUCCESS, message: 'Product updated', success: true, payload: data });
+        }
+    })
+}
+
 
 export const transferProductToStore = (req: Request, res: Response) => {
     const { productId, storeId } = req.body
